+++ src/vscode-checkbox-group/vscode-checkbox-group.styles.ts
@@ -1,0 +1,41 @@
+import {css, CSSResultGroup} from 'lit';
+import defaultStyles from '../includes/default.styles';
+import formHelperTextStyles from '../includes/formHelperTextStyles';
+
+const styles: CSSResultGroup = [
+  defaultStyles,
+  css`
+    :host {
+      display: block;
+    }
+
+    .wrapper {
+      display: flex;
+      flex-wrap: wrap;
+    }
+
+    :host([variant='vertical']) .wrapper {
+      display: block;
+    }
+
+    ::slotted(vscode-checkbox) {
+      margin-right: 20px;
+    }
+
+    ::slotted(vscode-checkbox:last-child) {
+      margin-right: 0;
+    }
+
+    :host([variant='vertical']) ::slotted(vscode-checkbox) {
+      display: block;
+      margin-bottom: 15px;
+    }
+
+    :host([variant='vertical']) ::slotted(vscode-checkbox:last-child) {
+      margin-bottom: 0;
+    }
+  `,
+  formHelperTextStyles,
+];
+
+export default styles;
